Refactor chat conversation component to improve modularity and readability.
